Organizing the models
Adding models directory
Inserting and creating models files- article.py, user.py, subject.py
Updating the imports in the migrations files.
